bridge_node/pyproject.toml
[tool.poetry]
name = "SovrynNbteBridge"
version = "0.0"
description = "Sovryn NBTE Bridge"
authors = ["Sovryn <info+sovryn@interjektio.fi>"]
license = "BSD"
readme = ["README.md"]
repository = "https://github.com/DistributedCollective/nbte-bridge"
classifiers = [
    "Programming Language :: Python",
    "Framework :: Pyramid",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]
keywords = ["web", "wsgi", "tet", "pyramid"]
packages = [{ include = "bridge" }]

[tool.poetry.plugins."paste.app_factory"]
main = "bridge:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
anemic = { git = "https://github.com/tetframework/anemic.git", rev = "3a397a5" }
python = "^3.11"
transaction = "*"
waitress = "*"
"zope.sqlalchemy" = "*"
SQLAlchemy = ">=2.0"
alembic = "*"
psycopg2 = "*"
click = "^7.0"
pyro5 = "^5.15"
venusian = "^3.1.0"
environ-config = "^23.2.0"
web3 = "^6.13.0"
eth-account = "^0.10.0"
eth-utils = "^2.3.1"
python-bitcointx = "^1.1.4"
pyramid = "^2.0.2"
python-dotenv = "^1.0.1"
pyord = "0.3.0"
pyramid-jinja2 = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
coverage = { extras = ["toml"], version = "^7.3.4" }
pytest-cov = "^4.1.0"
ruff = "^0.3.7"
pre-commit = "^3.6.0"
ipython = "^8.19.0"
ipdb = "^0.13.13"
pyyaml = "^6.0.1"

[tool.pytest.ini_options]
addopts = [
    # Enalbe coverage by default at some point if wanted
    #"--cov",
    "--cov-config",
    "pyproject.toml",
    "--cov-report",
    "lcov",
    "--cov-report",
    "term-missing",
]
testpaths = ["tests"]
python_files = ["test_*.py", "tests.py", "*_tests.py"]
markers = [
    "integration: marks tests as integration test, requiring startup of the slow integration testing harness (deselect with '-m \"not integration\"')",
]
filterwarnings = [
  "ignore:Secrets file not found.*:UserWarning:.*",
]

[tool.coverage.run]
branch = true
command_line = "-m pytest"
source = ["bridge"]
relative_files = true

[tool.coverage.report]
show_missing = true

[tool.coverage.lcov]
output = "coverage.lcov"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"
# Don't lint semi-autogenerated files, dependencies, tool scripts or the forked pyro client
# automatically.
extend-exclude = ["migrations", "submodules", "scripts", "bridge/common/p2p/client.py"]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "I",
  "S",
  "B",
  "A",
  "PLC",
  "N",
  "UP",
]

# N818: Don't care about exceptions that don't end with "Error"
# S101: Don't care about asserts for now
# N814: Don't care about differently cased imports, gives too much false positives
# S113: Can't determine good timeout for requests, so ignore for now. TODO: Maybe fix this
ignore = ["N818", "S101", "N814", "S113"]

[tool.ruff.lint.per-file-ignores]
# Don't care about security warnings or asserts in tests
"tests/*" = ["S"]
